Refactor events page files + remove create event for volunteers
