--- src/pages/Events.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import AdminEvents from '../components/events/AdminEvents/AdminEvents';
-
-function Events() {
-  return (
-    <div>
-      <AdminEvents />
-    </div>
-  );
-}
-
-export default Events;
