+++ src/pages/Events/Events.module.css
@@ -1,4 +1,4 @@
-@import '../../../common/global.css';
+@import '../../common/global.css';
 
 .events {
   margin: auto;
